recipes/rabbitsketch/requirement.txt
#python version=3.7
pymp-pypi==0.5.0
fastx==0.0.3
